Struct utp::UtpListener

pub struct UtpListener {
    // some fields omitted
}

A structure representing a socket server.

Examples

use utp::{UtpListener, UtpSocket};
use std::thread;

fn handle_client(socket: UtpSocket) {
    // ...
}

fn main() {
    // Create a listener
    let addr = "127.0.0.1:8080";
    let listener = UtpListener::bind(addr).unwrap();

    for connection in listener.incoming() {
        // Spawn a new handler for each new connection
        match connection {
            Ok((socket, _src)) => { thread::spawn(move || { handle_client(socket) }); },
            _ => ()
        }
    }
}

Methods

impl UtpListener

fn bind<A: ToSocketAddrs>(addr: A) -> Result<UtpListener>

Creates a new UtpListener bound to a specific address.

The resulting listener is ready for accepting connections.

The address type can be any implementor of the ToSocketAddr trait. See its documentation for concrete examples.

If more than one valid address is specified, only the first will be used.

fn accept(&self) -> Result<(UtpSocket, SocketAddr)>

Accepts a new incoming connection from this listener.

This function will block the caller until a new uTP connection is established. When established, the corresponding UtpSocket and the peer's remote address will be returned.

Notice that the resulting UtpSocket is bound to a different local port than the public listening port (which UtpListener holds). This may confuse the remote peer!

fn incoming(&self) -> Incoming

Returns an iterator over the connections being received by this listener.

The returned iterator will never return None.

fn local_addr(&self) -> Result<SocketAddr>

Returns the local socket address of this listener.