Struct utp::UtpStream

pub struct UtpStream {
    // some fields omitted
}

A structure that represents a uTP (Micro Transport Protocol) stream between a local socket and a remote socket.

The connection will be closed when the value is dropped (either explicitly or when it goes out of scope).

Examples

use utp::UtpStream;
use std::io::{Read, Write};

let mut stream = UtpStream::bind("127.0.0.1:1234").unwrap();
let _ = stream.write(&[1]);
let _ = stream.read(&mut [0; 1000]);

Methods

impl UtpStream

fn bind<A: ToSocketAddrs>(addr: A) -> Result<UtpStream>

Creates a uTP stream listening on the given address.

The address type can be any implementor of the ToSocketAddr trait. See its documentation for concrete examples.

If more than one valid address is specified, only the first will be used.

fn connect<A: ToSocketAddrs>(dst: A) -> Result<UtpStream>

Opens a uTP connection to a remote host by hostname or IP address.

The address type can be any implementor of the ToSocketAddr trait. See its documentation for concrete examples.

If more than one valid address is specified, only the first will be used.

fn close(&mut self) -> Result<()>

Gracefully closes connection to peer.

This method allows both peers to receive all packets still in flight.

fn local_addr(&self) -> Result<SocketAddr>

Returns the socket address of the local half of this uTP connection.