Struct utp::UtpSocket

pub struct UtpSocket {
    // some fields omitted
}

A uTP (Micro Transport Protocol) socket.

Methods

impl UtpSocket

fn bind<A: ToSocketAddrs>(addr: A) -> Result<UtpSocket>

Create a UTP socket from the given address. For now, I'll ignore all but the first address.

fn connect<A: ToSocketAddrs>(self, other: A) -> Result<UtpSocket>

Open a uTP connection to a remote host by hostname or IP address.

fn close(&mut self) -> Result<()>

Gracefully close connection to peer.

This method allows both peers to receive all packets still in flight.

fn recv_from(&mut self, buf: &mut [u8]) -> Result<(usize, SocketAddr)>

Receive data from socket.

On success, returns the number of bytes read and the sender's address. Returns Closed after receiving a FIN packet when the remaining inflight packets are consumed.

fn send_to(&mut self, buf: &[u8]) -> Result<usize>

Send data on socket to the remote peer. Returns nothing on success.