Struct utp::UtpSocket

pub struct UtpSocket {
    // some fields omitted
}

A structure that represents a uTP (Micro Transport Protocol) connection between a local socket and a remote socket.

The socket will be closed when the value is dropped (either explicitly or when it goes out of scope).

Examples

use utp::UtpSocket;

let mut socket = UtpSocket::bind("127.0.0.1:1234").unwrap();

let mut buf = [0; 1000];
let (amt, _src) = socket.recv_from(&mut buf).ok().unwrap();

let mut buf = &mut buf[..amt];
buf.reverse();
let _ = socket.send_to(buf).unwrap();

// Explicitly close the socket. You can either call `close` on the socket,
//explicitly drop it or just let it go out of scope.
socket.close();

Methods

impl UtpSocket

fn bind<A: ToSocketAddrs>(addr: A) -> Result<UtpSocket>

Creates a new UTP socket from the given address.

The address type can be any implementor of the ToSocketAddr trait. See its documentation for concrete examples.

If more than one valid address is specified, only the first will be used.

fn local_addr(&self) -> Result<SocketAddr>

Returns the socket address that this socket was created from.

fn connect<A: ToSocketAddrs>(other: A) -> Result<UtpSocket>

Opens a connection to a remote host by hostname or IP address.

The address type can be any implementor of the ToSocketAddr trait. See its documentation for concrete examples.

If more than one valid address is specified, only the first will be used.

fn close(&mut self) -> Result<()>

Gracefully closes connection to peer.

This method allows both peers to receive all packets still in flight.

fn recv_from(&mut self, buf: &mut [u8]) -> Result<(usize, SocketAddr)>

Receives data from socket.

On success, returns the number of bytes read and the sender's address. Returns 0 bytes read after receiving a FIN packet when the remaining inflight packets are consumed.

fn send_to(&mut self, buf: &[u8]) -> Result<usize>

Sends data on the socket to the remote peer. On success, returns the number of bytes written.

fn flush(&mut self) -> Result<()>

Consumes acknowledgements for every pending packet.